# Conflicts:
#	templates/vmbeez3/css/nature.css

+++ administrator/components/com_csvi/rantai/import/model.php
@@ -3,10 +3,10 @@
  * @package     CSVI
  * @subpackage  Imports
  *
- * @author      Roland Dalmulder <contact@csvimproved.com>
- * @copyright   Copyright (C) 2006 - 2016 RolandD Cyber Produksi. All rights reserved.
+ * @author      RolandD Cyber Produksi <contact@csvimproved.com>
+ * @copyright   Copyright (C) 2006 - 2018 RolandD Cyber Produksi. All rights reserved.
  * @license     GNU/GPLv3 http://www.gnu.org/licenses/gpl-3.0.html
- * @link        http://www.csvimproved.com
+ * @link        https://csvimproved.com
  */
 
 defined('_JEXEC') or die;
@@ -26,7 +26,7 @@ class RantaiImportModel extends RantaiModel
 	 * @var    CsviHelperImportFields
 	 * @since  6.0
 	 */
-	protected $fields = null;
+	protected $fields;
 
 	/**
 	 * Initialise the import.
@@ -35,9 +35,11 @@ class RantaiImportModel extends RantaiModel
 	 *
 	 * @return  bool  True on success | false on failure.
 	 *
-	 * @throws  CsviException
-	 *
 	 * @since   6.0
+	 *
+	 * @throws  Exception
+	 * @throws  CsviException
+	 * @throws  RuntimeException
 	 */
 	public function initialiseImport($csvi_process_id)
 	{
@@ -72,15 +74,44 @@ class RantaiImportModel extends RantaiModel
 				$this->loadTemplate($details->csvi_template_id);
 
 				// Setup the addon autoloader
-				$component = $this->template->get('component');
-				JLoader::registerPrefix(ucfirst($component), JPATH_ADMINISTRATOR . '/components/com_csvi/addon/' . $component);
+				$component      = $this->template->get('component');
+				$extension      = substr($component, 4);
+				$override       = $this->template->get('override');
+				$operation      = $this->template->get('operation');
+				$adminTemplate  = $this->getAdminTemplate();
+
+				if (!$component && !$operation)
+				{
+					throw new CsviException(JText::_('COM_CSVI_EXPORT_NO_COMPONENT_NO_OPERATION'), 514);
+				}
+
+				if ($override)
+				{
+					if (file_exists(JPATH_ADMINISTRATOR . '/templates/' . $adminTemplate . '/html/com_csvi/' . $component . '/model/import/' . $override . '.php'))
+					{
+						JLoader::registerPrefix(ucfirst($component), JPATH_ADMINISTRATOR . '/templates/' . $adminTemplate . '/html/com_csvi/' . $component);
+					}
+					else
+					{
+						JLoader::registerNamespace($extension, JPATH_ADMINISTRATOR . '/templates/' . $adminTemplate . '/html/com_csvi/');
+					}
+				}
+
+				// Setup the component autoloader
+				JLoader::registerNamespace($extension, JPATH_PLUGINS . '/csviaddon/');
+
+				// This loader is still needed for other classes than the import
+				JLoader::registerPrefix(ucfirst($component), JPATH_PLUGINS . '/csviaddon/' . $extension . '/' . $component);
 
 				// Setup the logger
 				$this->log->setActive($this->template->getLog());
 				$this->log->setLogId($details->csvi_log_id);
 
+				// Check the source
+				$source = $this->template->get('source', 'fromupload');
+
 				// Check if we already have a file to process
-				if (empty($details->processfile))
+				if (empty($details->processfile) && ($source !== 'fromdatabase'))
 				{
 					// Load the folder class
 					jimport('joomla.filesystem.folder');
@@ -94,39 +125,43 @@ class RantaiImportModel extends RantaiModel
 					}
 				}
 
-				// Set the file being processed
-				if ($details->processfile)
+				// Add a dummy file for the database as it doesn't use a real file
+				if ($source === 'fromdatabase')
 				{
-					// Set the folder to process
-					$this->processfolder = $details->processfolder;
-
-					// Set the file to process
-					$this->processfile = $details->processfile;
-
-					// Load the file
-					$this->loadImportFile();
-
-					// Tell the logger about the filename
-					$this->log->setFilename(basename($this->processfile));
-
-					// Load the fields
-					$this->loadFields();
-
-					// Both the file and fields have been setup, let's connect them
-					$this->file->setFields($this->fields);
-
-					// Let's setup the fields
-					$this->fields->setupFields();
-
-					// Move the file pointer if needed
-					if ($details->position > 0)
-					{
-						$this->file->setFilePos($details->position);
-					}
+					$details->processfile = 'database';
 				}
-				else
+
+				// Set the file being processed
+				if (!$details->processfile)
 				{
 					throw new CsviException(JText::_('COM_CSVI_NO_FOLDER_FOUND_TO_PROCESS'), 504);
+				}
+
+				// Set the folder to process
+				$this->processfolder = $details->processfolder;
+
+				// Set the file to process
+				$this->processfile = $details->processfile;
+
+				// Load the file
+				$this->loadImportFile();
+
+				// Tell the logger about the filename
+				$this->log->setFilename(basename($this->processfile));
+
+				// Load the fields
+				$this->loadFields();
+
+				// Both the file and fields have been setup, let's connect them
+				$this->file->setFields($this->fields);
+
+				// Let's setup the fields
+				$this->fields->setupFields();
+
+				// Move the file pointer if needed
+				if ($details->position > 0)
+				{
+					$this->file->setFilePos($details->position);
 				}
 
 				return true;
@@ -162,24 +197,29 @@ class RantaiImportModel extends RantaiModel
 			throw new CsviException(JText::_('COM_CSVI_NO_TEMPLATE_LOADED'), 501);
 		}
 
-		// Get the file parts
-		$upload_parts = pathinfo($this->processfile);
-
-		// Force an extension if needed
-		$force_ext = $this->template->get('use_file_extension');
-
-		if (!empty($force_ext))
+		if ($this->template->get('source') === 'fromdatabase')
 		{
-			$upload_parts['extension'] = $force_ext;
-		}
-
-		// Set the file helper
-		if (!array_key_exists('extension', $upload_parts))
-		{
-			throw new CsviException(JText::sprintf('COM_CSVI_NO_EXTENSION_FOUND_ON_IMPORT_FILE', $this->processfile), 502);
+			$fileclass = 'CsviHelperFileImportDatabase';
 		}
 		else
 		{
+			// Get the file parts
+			$upload_parts = pathinfo($this->processfile);
+
+			// Force an extension if needed
+			$force_ext = $this->template->get('use_file_extension');
+
+			if (!empty($force_ext))
+			{
+				$upload_parts['extension'] = $force_ext;
+			}
+
+			// Set the file helper
+			if (!array_key_exists('extension', $upload_parts))
+			{
+				throw new CsviException(JText::sprintf('COM_CSVI_NO_EXTENSION_FOUND_ON_IMPORT_FILE', $this->processfile), 502);
+			}
+
 			$fileclass = 'CsviHelperFileImport';
 
 			switch (strtolower($upload_parts['extension']))
@@ -251,6 +291,12 @@ class RantaiImportModel extends RantaiModel
 			{
 				$this->helper->unpublishBeforeImport($this->template, $this->log, $this->db);
 			}
+
+			// Delete rows before import
+			if ($this->helper && method_exists($this->helper, 'truncateTableBeforeImport'))
+			{
+				$this->helper->truncateTableBeforeImport();
+			}
 		}
 	}
 
@@ -260,23 +306,65 @@ class RantaiImportModel extends RantaiModel
 	 * @return  void.
 	 *
 	 * @since   6.0
+	 *
+	 * @throws  UnexpectedValueException
 	 */
 	public function onAfterImport()
 	{
 		// Check if the import file needs to be deleted
-		if ($this->template->get('delete_file') && $this->template->get('source', 'fromupload') == 'fromserver')
-		{
-			// Get the template location
-			$location = $local_file = JPath::clean($this->template->get('local_csv_file'), '/');
+		$from = $this->template->get('source', 'fromupload');
 
-			if (is_dir($location))
+		if ($this->template->get('delete_file') && in_array($from, array('fromserver', 'fromftp'), true))
+		{
+			switch ($from)
 			{
-				JFile::delete($location . '/' . basename($this->processfile));
-			}
-			else
-			{
-				$this->log->add('Process file:' . $this->processfile);
-				JFile::delete($this->processfile);
+				case 'fromftp':
+					$ftpFile = $this->template->get('ftpfile');
+
+					// Start the FTP
+					jimport('joomla.client.ftp');
+					$ftp = JClientFtp::getInstance(
+						$this->template->get('ftphost'),
+						$this->template->get('ftpport'),
+						array(),
+						$this->template->get('ftpusername'),
+						$this->template->get('ftppass')
+					);
+
+					$ftpRoot = $this->template->get('ftproot', '');
+
+					if (substr($ftpRoot, -1) !== '/')
+					{
+						$ftpRoot .= '/';
+					}
+
+					$serverFile = $ftpRoot . JPath::clean($ftpFile, '/');
+
+					// Check if a specific file has been set
+					if (strlen($ftpFile) === 0)
+					{
+						$serverFile = $ftpRoot . basename($this->processfile);
+					}
+
+					$this->log->add('Delete file:' . $serverFile);
+					$ftp->delete($serverFile);
+					break;
+				case 'fromserver':
+					$local_file = JPath::clean($this->template->get('local_csv_file'), '/');
+
+					if (is_dir($local_file))
+					{
+						JFile::delete($local_file . '/' . basename($this->processfile));
+					}
+					else
+					{
+						$this->log->add('Delete temporary file:' . $this->processfile);
+						JFile::delete($this->processfile);
+
+						$this->log->add('Delete original file:' . $local_file);
+						JFile::delete($local_file);
+					}
+					break;
 			}
 		}
 	}
@@ -333,6 +421,7 @@ class RantaiImportModel extends RantaiModel
 						{
 							case 'ftpusername':
 							case 'ftppass':
+							case 'database_password':
 								break;
 							case 'source':
 								$this->log->add(JText::_('COM_CSVI_JFORM_' . $name . '_LABEL') . ': ' . $value, false);
@@ -366,8 +455,6 @@ class RantaiImportModel extends RantaiModel
 	 *
 	 * @param   bool  $isCli  Set if the import is run from CLI
 	 *
-	 * @todo    Routine overrides
-	 *
 	 * @return  bool  true if we continue importing | false if the import is finished/terminated.
 	 *
 	 * @throws  CsviException
@@ -376,6 +463,12 @@ class RantaiImportModel extends RantaiModel
 	 */
 	final public function runImport($isCli = false)
 	{
+		// Load the template settings
+		$component = $this->template->get('component');
+		$extension = substr($component, 4);
+		$operation = $this->template->get('operation');
+		$override  = $this->template->get('override');
+
 		// Set the system limits to the user settings if needed
 		$this->systemLimits();
 
@@ -386,7 +479,26 @@ class RantaiImportModel extends RantaiModel
 		}
 
 		// Load the import routine
-		$classname = ucwords($this->template->get('component')) . 'ModelImport' . ucwords($this->template->get('operation'));
+		$classname = '\\' . $extension . '\\' . $component . '\\model\import\\' . $operation;
+
+		// Load the import routine
+		$adminTemplate = $this->getAdminTemplate();
+
+		if ($override)
+		{
+			$classname = '\\' . $extension . '\\' . $component . '\\model\import\\' . $override;
+
+			// Check for the old style override pre 7.2.0
+			if (file_exists(JPATH_ADMINISTRATOR . '/templates/' . $adminTemplate . '/html/com_csvi/' . $component . '/model/import/' . $override . '.php'))
+			{
+				$classname = ucwords($component) . 'ModelImport' . ucwords($override);
+
+				if (!class_exists($classname))
+				{
+					$classname = '\\' . $extension . '\\' . $component . '\\model\import\\' . $override;
+				}
+			}
+		}
 
 		// Instantiate the import routine
 		$routine = new $classname(
@@ -407,8 +519,7 @@ class RantaiImportModel extends RantaiModel
 				->update($this->db->quoteName('#__csvi_templates'))
 				->set($this->db->quoteName('lastrun') . ' = ' . $this->db->quote(JFactory::getDate()->toSql()))
 				->where($this->db->quoteName('csvi_template_id') . ' = ' . (int) $this->template->getId());
-			$this->db->setQuery($query);
-			$this->db->execute();
+			$this->db->setQuery($query)->execute();
 
 			// Initialize process data so the import starts
 			$processdata = true;
@@ -436,11 +547,13 @@ class RantaiImportModel extends RantaiModel
 				}
 				else
 				{
-					$nolines = $this->template->get('import_nolines');
+					$nolines = $this->template->get('import_nolines', 1000);
 				}
 
 				if (($this->recordsProcessed + 1) <= $nolines)
 				{
+					$this->fields->setProcessRecord(true);
+
 					// Load the data
 					$result = $this->file->readNextLine();
 
@@ -507,6 +620,13 @@ class RantaiImportModel extends RantaiModel
 							$return = false;
 						}
 					}
+
+					// Check for time between imports
+					if ($this->recordsProcessed === (int) $nolines && !$isCli)
+					{
+						$importWait = $this->template->get('import_wait', 5);
+						sleep($importWait);
+					}
 				}
 				else
 				{
@@ -549,24 +669,27 @@ class RantaiImportModel extends RantaiModel
 	 *
 	 * @since   3.0
 	 */
-	private function finishProcess($finished=false)
+	private function finishProcess($finished = false)
 	{
 		// Check if the import is finished or if we are going to reload
 		if ($finished)
 		{
-			// Close the file
-			$this->file->closeFile(false);
-
-			// Remove the file that has just been imported
-			if (JFile::delete($this->file->getFilename()))
+			if ($this->template->get('source') !== 'fromdatabase')
 			{
-				// Remove the processfile setting, this is needed for folder processing to continue
-				$query = $this->db->getQuery(true)
-					->update($this->db->quoteName('#__csvi_processes'))
-					->set($this->db->quoteName('processfile') . ' = ""')
-					->set($this->db->quoteName('position') . ' = 0')
-					->where($this->db->quoteName('csvi_process_id') . ' = ' . (int) $this->runId);
-				$this->db->setQuery($query)->execute();
+				// Close the file
+				$this->file->closeFile(false);
+
+				// Remove the file that has just been imported
+				if (JFile::delete($this->file->getFilename()))
+				{
+					// Remove the processfile setting, this is needed for folder processing to continue
+					$query = $this->db->getQuery(true)
+						->update($this->db->quoteName('#__csvi_processes'))
+						->set($this->db->quoteName('processfile') . ' = ""')
+						->set($this->db->quoteName('position') . ' = 0')
+						->where($this->db->quoteName('csvi_process_id') . ' = ' . (int) $this->runId);
+					$this->db->setQuery($query)->execute();
+				}
 			}
 
 			// Set the log end timestamp
